+++ core/server/src/main/java/alluxio/worker/DefaultAlluxioWorker.java
@@ -101,45 +101,46 @@ public final class DefaultAlluxioWorker implements AlluxioWorkerService {
 
   /**
    * Creates a new instance of {@link DefaultAlluxioWorker}.
-   *
-   * @throws Exception if the creation fails
    */
-  public DefaultAlluxioWorker() throws Exception {
-    mWorkerId = new AtomicReference<>();
-    mStartTimeMs = System.currentTimeMillis();
-    mBlockWorker = new DefaultBlockWorker(mWorkerId);
-    mFileSystemWorker = new DefaultFileSystemWorker(mBlockWorker, mWorkerId);
+  public DefaultAlluxioWorker() {
+    try {
+      mWorkerId = new AtomicReference<>();
+      mStartTimeMs = System.currentTimeMillis();
+      mBlockWorker = new DefaultBlockWorker(mWorkerId);
+      mFileSystemWorker = new DefaultFileSystemWorker(mBlockWorker, mWorkerId);
 
-    mAdditionalWorkers = new ArrayList<>();
-    List<? extends Worker> workers = Lists.newArrayList(mBlockWorker, mFileSystemWorker);
-    // Discover and register the available factories
-    // NOTE: ClassLoader is explicitly specified so we don't need to set ContextClassLoader
-    ServiceLoader<WorkerFactory> discoveredMasterFactories =
-        ServiceLoader.load(WorkerFactory.class, WorkerFactory.class.getClassLoader());
-    for (WorkerFactory factory : discoveredMasterFactories) {
-      Worker worker = factory.create(workers);
-      if (worker != null) {
-        mAdditionalWorkers.add(worker);
+      mAdditionalWorkers = new ArrayList<>();
+      List<? extends Worker> workers = Lists.newArrayList(mBlockWorker, mFileSystemWorker);
+      // Discover and register the available factories
+      // NOTE: ClassLoader is explicitly specified so we don't need to set ContextClassLoader
+      ServiceLoader<WorkerFactory> discoveredWorkerFactories =
+          ServiceLoader.load(WorkerFactory.class, WorkerFactory.class.getClassLoader());
+      for (WorkerFactory factory : discoveredWorkerFactories) {
+        Worker worker = factory.create(workers);
+        if (worker != null) {
+          mAdditionalWorkers.add(worker);
+        }
       }
+
+      // Setup web server
+      mWebServer =
+          new WorkerWebServer(NetworkAddressUtils.getBindAddress(ServiceType.WORKER_WEB), this,
+              mBlockWorker, NetworkAddressUtils.getConnectHost(ServiceType.WORKER_RPC), mStartTimeMs);
+
+      // Setup Thrift server
+      mTransportProvider = TransportProvider.Factory.create();
+      mThriftServerSocket = createThriftServerSocket();
+      int rpcPort = NetworkAddressUtils.getThriftPort(mThriftServerSocket);
+      String rpcHost = NetworkAddressUtils.getThriftSocket(mThriftServerSocket).getInetAddress().getHostAddress();
+      mRpcAddress = new InetSocketAddress(rpcHost, rpcPort);
+      mThriftServer = createThriftServer();
+
+      // Setup Data server
+      mDataServer = DataServer.Factory
+          .create(NetworkAddressUtils.getBindAddress(ServiceType.WORKER_DATA), this);
+    } catch (Exception e) {
+      Throwables.propagate(e);
     }
-
-    // Setup web server
-    mWebServer =
-        new WorkerWebServer(NetworkAddressUtils.getBindAddress(ServiceType.WORKER_WEB), this,
-            mBlockWorker, NetworkAddressUtils.getConnectHost(ServiceType.WORKER_RPC), mStartTimeMs);
-
-    // Setup Thrift server
-    mTransportProvider = TransportProvider.Factory.create();
-    mThriftServerSocket = createThriftServerSocket();
-    int rpcPort = NetworkAddressUtils.getThriftPort(mThriftServerSocket);
-    String rpcHost =
-        NetworkAddressUtils.getThriftSocket(mThriftServerSocket).getInetAddress().getHostAddress();
-    mRpcAddress = new InetSocketAddress(rpcHost, rpcPort);
-    mThriftServer = createThriftServer();
-
-    // Setup Data server
-    mDataServer = DataServer.Factory
-        .create(NetworkAddressUtils.getBindAddress(ServiceType.WORKER_DATA), this);
   }
 
   @Override
